pyproject.toml
[project]
name = "bci-interface"
version = "0.1.0"
description = "Brain-Computer Interface for binary communication"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "websockets>=10.0",
    "brainflow>=5.0",
    "numpy>=1.21",
    "python-multipart>=0.0.5",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[dependency-groups]
dev = [
    "black>=22.3.0",
    "isort>=5.10.1",
    "mypy>=0.950",
    "pytest>=7.1.2",
]

[project.optional-dependencies]
dev = [
    "black>=22.3.0",
    "isort>=5.10.1",
    "mypy>=0.950",
    "pytest>=7.1.2",
]

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true
